+++ .bandit
@@ -1,3 +1,3 @@
 [bandit]
 targets: src/
-exclude: src/shipchain_common/test_utils.py
+exclude: src/shipchain_common/test_utils/
